requirements.txt
numpy
fastapi
uvicorn
pandas
matplotlib
path.py
aiofiles
spiper
sklearn
eutils>=0.5,<0.6
xmltodict
pydantic
spiper@https://github.com/shouldsee/spiper/tarball/0.1.5==0.1.5
attrdict
